update the docs to reflect the decision to not persist cluster state
Instead of persisting information about the state of the ceph clusters
created the service should instead just perform tasks and report back
on the status them. The state of the cluster is better acquired
by asking the cluster itself and keeping everything up-to-date in this
service would be difficult and unnecessary.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>