util: set fetch_directory to $HOME/fetch/
ceph-ansible tries to write to fetch_directory. It defaults this
variable to a location in the current working directory, and when we're
running ansible in prodution, we don't want to grant the UID write
access to cwd, since that is /usr/share/ceph-ansible.

Set fetch_directory to a location inside $HOME, since we know that will
be writable.