set monitor_name when host is a mon and osd in osd/configure/
The var monitor_name is set in the ceph-mon role and is needed to
restart the monitors. Because we're configuring an osd that is also a
mon, we need to restart the monitor process as well. Without using the
ceph-mon role we manually set monitor_name so that the process
will restart correctly.

Signed-off-by: Andrew Schoen <aschoen@redhat.com>