util: add trailing newline to inventory files
In simple cluster cases (like a single monitor), this inventory file may
only have one host. If it's the last line in the file, `cat` puts it on
the same line as the prompt... which makes my eyes skip over it,
thinking it's empty.

Add a trailing newline so it's easier to quickly read the file with
`cat`.