/setup/key/: don't run ssh-keygen during HTTP request
Commit f8deeae628748be69e47cc23ace4f951a960fc9e changed the
ceph-installer application to run `ssh-keygen` as part of the start up
process. If `ssh-keygen` fails, then the app will fail to start (raising
a RuntimeError).

In other words, we can trust that the SSH public key file will always be
available when a client hits the /setup/key/ endpoint. If the file is
missing at this point, we should treat that as a hard error condition
(logging on the server, returning an HTTP error to the client), and
never try to generate an SSH key within the gunicorn child worker
process.

The purpose of this change is to further harden the code against races
during /setup/key/, where more than one worker could possibly run
`ssh-keygen` at the same time.