setup.py: add "bump" command
This command will increment the last number in __version__
automatically and commit all changes to Git with a standard commit
message.

This uses our new read_module_contents() helper from the previous
commit.

(In the vast majority of cases, we simply want to bump the final version
number part, so this optimizes for that scenario to make that easier. If
we want to go to "2.0.0" or "3.0.0", that will have to be done by hand,
or else add more user_options to this command to make it more flexible.)