selinux: set ceph_installer_t to permissive
We have still found various AVC denials when running Skyring (Tendrl)
with SELinux in enforcing mode. To make sure ceph-installer works while
we fix the remaining issues with the policies, we are going to set the
entire ceph_installer_t domain to permissive.

With this change, the ceph-installer service still runs confined, but
AVC denials will not prevent the service from operating.

See https://lwn.net/Articles/303216/ for background on permissive
SELinux domains.