add makefile for building RPMs from Git
Update the RPM packaging so that we can dynamically generate RPMs with
"make rpm".

The RPM .spec file becomes a template, we get rid of the %{commitdate}
convention and move to a "git describe" convention, similar to what
ceph.git uses.

The purpose of this change is to set up a simple command to allow
Jenkins to create RPMs on the fly.

For now this Makefile hardcodes a custom EPEL 7-based mock chroot from
copr (ktdreyer-ceph-installer). EPEL 7 itself lacks a couple RPM
dependencies, so I had to use this separate repository. We could improve
this later, possibly by moving this project into the CentOS CI
infrastructure.